requirements.txt
gpxpy>=1.5.0
numpy>=1.23
scipy>=1.9
Pillow>=9.0
pyproj>=3.4
requests>=2.28
trimesh>=3.18
shapely>=2.0
streamlit>=1.28
python-dotenv>=1.0
networkx>=3.0
mapbox-earcut>=1.0
